Move interfaces tests to separate file
